fix/issue-with-heap-corruption-is-fixed
The problem is due to `*ptr = value` which reads and writes. Reading part is UB as the memory is uninitialized. However, we do not need to read. The problem is fixed by using `std::ptr::write` instead.

The problem had not been observed before due to lack of tests with heap allocated types. Test coverage is increased by adding String tests.